# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "version": "1.0.0",
    "description": "A comprehensive blog API with authentication, CRUD operations, likes, and comments",
    "contact": {
      "name": "API Support",
      "email": "support@blogapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "Blog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "My First Blog Post"
          },
          "content": {
            "type": "string",
            "example": "This is the content of my blog post..."
          },
          "excerpt": {
            "type": "string",
            "example": "A brief summary of the blog post"
          },
          "author_id": {
            "type": "integer",
            "example": 1
          },
          "is_published": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "author_username": {
            "type": "string",
            "example": "johndoe"
          },
          "author_full_name": {
            "type": "string",
            "example": "John Doe"
          },
          "likes_count": {
            "type": "integer",
            "example": 5
          },
          "comments_count": {
            "type": "integer",
            "example": 3
          },
          "is_liked": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 10
              },
              "total": {
                "type": "integer",
                "example": 100
              },
              "totalPages": {
                "type": "integer",
                "example": 10
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/blogs": {
      "post": {
        "summary": "Create a new blog post",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "My First Blog Post"
                  },
                  "content": {
                    "type": "string",
                    "example": "This is the content of my blog post..."
                  },
                  "excerpt": {
                    "type": "string",
                    "example": "A brief summary"
                  },
                  "is_published": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Blog created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Blog"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all published blogs (public)",
        "tags": [
          "Blogs"
        ],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of blogs per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Blogs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/blogs/{id}": {
      "get": {
        "summary": "Get a specific blog by ID",
        "tags": [
          "Blogs"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Blog ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Blog retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Blog retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Blog"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Blog not found"
          }
        }
      },
      "put": {
        "summary": "Update a blog post",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Blog ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Updated Blog Title"
                  },
                  "content": {
                    "type": "string",
                    "example": "Updated content..."
                  },
                  "excerpt": {
                    "type": "string",
                    "example": "Updated excerpt"
                  },
                  "is_published": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog updated successfully"
          },
          "403": {
            "description": "Not authorized to update this blog"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a blog post",
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Blog ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted successfully"
          },
          "403": {
            "description": "Not authorized to delete this blog"
          },
          "404": {
            "description": "Blog not found"
          }
        }
      }
    },
...
  },
...
}